# Frontend/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Backend",
    "version": "1.0"
  },
  "paths": {
    "/api/Damage": {
      "get": {
        "tags": [
          "Damage"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Damage"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Damage"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Damage"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Damage"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Damage"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Damage"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Damage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Damage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Damage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Damage"
                }
              }
            }
          }
        }
      }
    },
    "/api/Damage/{id}": {
      "get": {
        "tags": [
          "Damage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Damage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Damage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Damage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Damage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Damage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Damage"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Damage"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Damage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Damage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/DamageType"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          }
        },
        "additionalProperties": false
      },
      "DamageType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Location": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "nullable": true
          },
          "streetNumber": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}